test:update test for pkg/event and pkg/event/schemas
